third_party/WebKit/LayoutTests/editing/pasteboard/no-clipboard-event-to-hidden-selection.html

<!doctype html>
<title>A hidden carret should not receive ClipboardEvent</title>
<link rel="help" href="https://www.w3.org/TR/clipboard-apis/#fire-a-clipboard-event">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<input type="checkbox" id="checkbox">
<textarea id="textcontrol">HEY</textarea>

<script>
testClipboardEvent = function(eventName, selectedElement) {
  let selection_hit_count = 0;
  let document_hit_count = 0;
  let last_target;
  selectionLog = function(e) {
    selection_hit_count++;
    last_target = e.target;
  }
  documentLog = function(e) {
    document_hit_count++;
    last_target = e.target;
  }
  selectedElement.addEventListener(eventName, selectionLog);
  document.addEventListener(eventName, documentLog);

  selectedElement.focus();
  testRunner.execCommand(eventName);
  assert_equals(selection_hit_count, 1, "The <textarea> should receive one " + eventName + "-event.")
  assert_equals(document_hit_count, 1, "The document should receive one bubbled " + eventName + "-event.")
  assert_equals(last_target, selectedElement, "The <textarea> should be the " + eventName + "-event's target.")

  checkbox.focus();
  testRunner.execCommand(eventName);
  assert_equals(selection_hit_count, 1, "The <textarea> is hidden so it should not receive the " + eventName + "-events.")
  assert_equals(document_hit_count, 2, "The document should receive one " + eventName + "-event.")
  assert_equals(last_target, document.body, "The document should be the " + eventName + "-event's target.")

  checkbox.blur();
  testRunner.execCommand(eventName);
  assert_equals(selection_hit_count, 1, "The <textarea> is hidden so it should not receive the " + eventName + "-events.")
  assert_equals(document_hit_count, 3, "The document should receive one " + eventName + "-event.")
  assert_equals(last_target, document.body, "The document should be the " + eventName + "-event's target.")

  document.body.focus();
  testRunner.execCommand(eventName);
  assert_equals(selection_hit_count, 1, "The <textarea> is hidden so it should not receive the " + eventName + "-events.")
  assert_equals(document_hit_count, 4, "The document should receive another " + eventName + "-event.")
  assert_equals(last_target, document.body, "The document should be the " + eventName + "-event's target.")
}

test(() => testClipboardEvent('copy', textcontrol),  'Hidden carrets in textcontrols are not targeted by copy-events.');
test(() => testClipboardEvent('paste', textcontrol), 'Hidden carrets in textcontrols are not targeted by paste-events.');
test(() => testClipboardEvent('cut', textcontrol),   'Hidden carrets in textcontrols are not targeted by cut-events.');
</script>
